+++ src/Helpers/CustomPathGenerator.php
@@ -2,7 +2,7 @@
 
 namespace Larrock\Core\Helpers;
 
-use Spatie\MediaLibrary\Media;
+use Spatie\MediaLibrary\Models\Media;
 use Spatie\MediaLibrary\PathGenerator\PathGenerator;
 
 /**
@@ -13,32 +13,35 @@ use Spatie\MediaLibrary\PathGenerator\PathGenerator;
  */
 class CustomPathGenerator implements PathGenerator
 {
-	/**
-	 * Get the path for the given media, relative to the root storage path.
-	 *
-	 * @param \Spatie\MediaLibrary\Media $media
-	 *
-	 * @return string
-	 */
-    public function getPath(Media $media) : string
-	{
-		$model = class_basename($media->model_type);
-		return $model .'/';
-	}
-	/**
-	 * Get the path for conversions of the given media, relative to the root storage path.
-	 *
-	 * @param \Spatie\MediaLibrary\Media $media
-	 *
-	 * @return string
-	 */
-    public function getPathForConversions(Media $media) : string
-	{
-		return class_basename($media->model_type). '/'. $media->name .'/';
-	}
+    /*
+     * Get the path for the given media, relative to the root storage path.
+     */
+    public function getPath(Media $media): string
+    {
+        return $this->getBasePath($media).'/';
+    }
 
+    /*
+     * Get the path for conversions of the given media, relative to the root storage path.
+     */
+    public function getPathForConversions(Media $media): string
+    {
+        return $this->getBasePath($media).'/conversions/';
+    }
+
+    /*
+     * Get the path for responsive images of the given media, relative to the root storage path.
+     */
     public function getPathForResponsiveImages(Media $media): string
     {
-        return class_basename($media->model_type). '/'. $media->name .'/r/';
+        return $this->getBasePath($media).'/responsive-images/';
+    }
+
+    /*
+     * Get a unique base path for the given media.
+     */
+    protected function getBasePath(Media $media): string
+    {
+        return class_basename($media->model_type) .'/'. $media->name;
     }
 }